#add new func of find max, min, avg size of files in dir

+++ log.sh
@@ -10,8 +10,18 @@ threshold=${2:-70}
 num_archive=${3:-5}
 backup_dir='/BACKUP'
 
-if [ ! -d "$1" ]; then
-    echo "[LOG] Error: directory was not requests. Hint: ./log.sh /Name_of_directiry"
+if [ -z "$log_dir" ]; then
+    echo "[LOG] Error: Directory path not provided. Use ./log.sh /directory_name ."
+    exit 1
+fi
+
+if [ ! -d "$log_dir" ]; then
+    echo "[LOG] Error: Directory $log_dir does not exist."
+    exit 1
+fi
+
+if [[ $threshold -lt 1 || $threshold -gt 100 ]]; then
+    echo "[LOG] Error: Invalid threshold value. It must beetwen 1 and 100."
     exit 1
 fi
 
@@ -23,6 +33,25 @@ if [ $usage -ge $threshold ]; then
     mkdir -p "$backup_dir"
 
     files=$(ls -1t $log_dir | head -n $num_archive)
+
+    all_count_files=$(ls -1A $log_dir | grep -v 'lost+found' | wc -l)
+    file_paths=""
+    for file in $(ls -1t $log_dir | head -n $all_count_files); do
+        if [ -d "$log_dir/$file" ]; then
+            continue
+        fi
+        file_paths="$file_paths $log_dir/$file"
+    done
+    file_sizes=$(stat --format="%s" $file_paths)
+    
+    total_size=$(echo "$file_sizes" | awk '{s+=$1} END {print s}')
+    max_size_file=$(echo "$file_sizes" | sort -n | tail -1)
+    min_size_file=$(echo "$file_sizes" | sort -n | head -1)
+    avg_size_file=$(echo "$total_size/$all_count_files" | bc)
+    
+    echo "[LOG] Max size file: $max_size_file bytes"
+    echo "[LOG] Min size file: $min_size_file bytes"
+    echo "[LOG] Avg size file: $avg_size_file bytes"
     
     if [ -z "$files" ]; then
         echo "[LOG] ERROR: Files not found in folder $log_dir"
@@ -30,8 +59,12 @@ if [ $usage -ge $threshold ]; then
     fi
 
     sudo tar -czf $backup_dir/backup_$(date +%Y%m%d%_H%M%S).tar.gz -C $log_dir $files
+    echo "[LOG] Files was archived successfully"
 
     for file in $files; do
+        if [ -d "$log_dir/$file" ]; then
+            continue
+        fi
         echo "[LOG] Removing file $file from $log_dir"
         sudo rm -f $log_dir/$file
     done
